webpt_qa/requirements.txt
selenium==4.23.1
webdriver-manager==4.0.2
PyYAML==6.0.2
python-dotenv==1.0.1
rich==13.7.1